backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
python-dotenv==1.0.1
apscheduler==3.11.2
playwright==1.44.0
